requirements.txt
Flask
gspread
google-auth
googletrans==4.0.0-rc1
python-dotenv
